type changed
git-svn-id: https://svn.urbansim.org/src/trunk@20111 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ psrc_parcel/emme/travel_model_output.py
@@ -30,8 +30,8 @@ class TravelModelOutput(ParentTravelModelOutput):
         in_storage.write_table(
                 table_name=table_name,
                 table_data={
-                    'from_zone_id':zone_set.get_id_attribute()[comb_index[0].ravel()],
-                    'to_zone_id':zone_set.get_id_attribute()[comb_index[1].ravel()],
+                    'from_zone_id':zone_set.get_id_attribute()[comb_index[0].ravel()].astype('int32'),
+                    'to_zone_id':zone_set.get_id_attribute()[comb_index[1].ravel()].astype('int32'),
                     }
             )
